requirements.txt
aws-cdk.aws-route53==1.74.0
aws-cdk.aws-ssm==1.74.0
aws-cdk.aws-certificatemanager==1.74.0
aws-cdk.aws-cognito==1.74.0
aws-cdk-core==1.74.0
